requirements.txt
requests
beautifulsoup4
pandas
streamlit
SQLAlchemy
pymysql
python-dotenv

langchain-openai
openai

playwright
playwright-stealth